+++ spec/models/concerns/balance_spec.rb
@@ -1,7 +1,6 @@
 require 'spec_helper'
 
-describe Spree::Order do
-
+describe Balance do
   context "#outstanding_balance" do
     context 'when orders are in cart state' do
       let(:order) { build(:order, total: 100, payment_total: 10, state: 'cart') }
